pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
packages = [{ include = "app" }]
description = ""
authors = ["abdulazizmirjalilov <abdulaziz.mirjalilov@ventionteams.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
alembic = "^1.14.1"
fastapi = "^0.115.8"
pydantic = { extras = ["email"], version = "^2.10.6" }
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.38"
asyncpg = "^0.30.0"
celery = "^5.4.0"
redis = "^5.2.1"
flower = "^2.0.1"
pytest = "^8.3.4"
pydantic-settings = "^2.7.1"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
sib-api-v3-sdk = "^7.6.0"
jinja2 = "^3.1.5"
aiohttp = "^3.11.13"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pre-commit = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311', 'py311', 'py313']
indent-width = 4

[tool.ruff]
target-version = "py313"
line-length = 88
indent-width = 4
fix = true

select = [
    # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "F", # Pyflakes
    # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "E", # pycodestyle
    "W", # Warning
    # https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    # https://docs.astral.sh/ruff/rules/#mccabe-c90
    "C", # Complexity (mccabe+) & comprehensions
    # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "UP", # pyupgrade
    # https://docs.astral.sh/ruff/rules/#isort-i
    "I", # isort
]
ignore = [
    # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "E402", # module level import not at top of file
    # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "UP006", # use-pep585-annotation
    "UP007", # use-pep604-annotation
    "E741", # Ambiguous variable name
    # "UP035", # deprecated-assertion
]
[tool.ruff.per-file-ignores]
"__init__.py" = [
    "F401", # unused import
    "F403", # star imports
]

[tool.ruff.mccabe]
max-complexity = 24

[tool.ruff.pydocstyle]
convention = "numpy"
